#69 [Refactor] 아파트일정 어뎁터 분리
