+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/entiremenu/AptOperationInfo/fragment/OperationSecondFragment.kt
@@ -6,12 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView.LayoutManager
 import com.google.android.material.divider.MaterialDividerItemDecoration
 import kr.co.lion.application.finalproject_aparttalk.R
 import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentOperationSecondBinding
 import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.OperationInfoActivity
-import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.adapter.OperationSecondAdapter
+import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptOperationInfo.adapter.OperationSecondRecyclerView
 import kr.co.lion.application.finalproject_aparttalk.util.AptOperationInfoFragmentName
 
 class OperationSecondFragment : Fragment() {
@@ -49,7 +48,7 @@ class OperationSecondFragment : Fragment() {
         fragmentOperationSecondBinding.apply {
             recyclerOperationInfoList.apply {
                 // 어댑터 설정
-                adapter = OperationSecondAdapter(childFragmentManager)
+                adapter = OperationSecondRecyclerView(childFragmentManager)
                 // 레이아웃
                 layoutManager = LinearLayoutManager(requireContext())
                 // 데코
