+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/entiremenu/AptSchedule/fragment/AptScheduleShowBottomSheetFragment.kt
@@ -1,17 +1,15 @@
-package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptSchedule
+package kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptSchedule.fragment
 
 import android.app.Dialog
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.media3.common.util.Log
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import kr.co.lion.application.finalproject_aparttalk.R
 import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentAptScheduleShowBottomSheetBinding
+import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.AptSchedule.AptScheduleActivity
 
 class AptScheduleShowBottomSheetFragment : BottomSheetDialogFragment() {
