+++ app/src/main/res/layout/fragment_apt_schedule_show_bottom_sheet.xml
@@ -7,7 +7,7 @@
     android:background="@drawable/bottom_sheet_rounding"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context=".ui.entiremenu.AptSchedule.AptScheduleShowBottomSheetFragment">
+    tools:context=".ui.entiremenu.AptSchedule.fragment.AptScheduleShowBottomSheetFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
